♻️ move settings page builder tests to a dedicated tests file
